requirements.txt
about-time==3.1.1
alive-progress==2.4.1
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
cronitor==4.5.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
fonttools==4.33.3
grapheme==0.6.0
greenlet==1.1.2
humanize==4.2.0
idna==3.3
ipykernel==6.13.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.10.0
kiwisolver==1.4.3
lxml==4.8.0
matplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.3
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.12.0
pymongo==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
pyzmq==23.1.0
requests==2.27.1
schedule==1.1.0
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
stack-data==0.2.0
toml==0.10.2
tornado==6.1
traitlets==5.2.2.post1
urllib3==1.26.9
wcwidth==0.2.5
